requirements.txt
fair-esm>=2.0.0
torch>=2.0.0
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
requests>=2.31.0
biopython>=1.81
tqdm>=4.65.0
networkx>=3.1
pyvis>=0.3.2
